removed debugging features, preparing for profile page update
homepage no longer alerts for every game in the database.
homepage now truncates text that would otherwise ruin the formatting of
the result boxes. this may turn into a "Read More ->" button in the
future.
profile page is currently not functional, but will show the useful bits
of the check out table. this table format will be changed for better
formatting options.